Consolidate database column builders
