+++ webapp/AGENTS.md
@@ -62,7 +62,7 @@
   - The server-only guarded `src/db/index.ts` entry point exports the Promise Drizzle handle and the native Effect database service; import both through `@/db`. Share its single application `pg.Pool` with `drizzle-orm/effect-postgres` through `PgClient.fromPool` and one `ManagedRuntime`, while keeping Better Auth on the Promise handle and native Effect workflows behind the one framework bridge. Keep node-postgres pool lifecycle defaults; do not add `allowExitOnIdle`, a parallel pool, global PostgreSQL type-parser table, or Nitro/signal lifecycle plumbing. The migration runner's dedicated postgres.js sessions remain the intentional exception.
   - `schema.server.ts` contains the drizzle schema
   - Keep generally reusable database primitives in `src/db/lib`; keep table-specific domain interfaces and the migration runner directly under `src/db`.
-  - Define shared PostgreSQL columns with `caseInsensitiveText()`, `timestampWithTimeZone()`, `timestamps()`, `lockVersion()`, and `uuidV7PrimaryKey()` from `src/db/lib/columns.server.ts`; do not inline equivalent Drizzle builders.
+  - Define shared PostgreSQL columns with `caseInsensitiveText()`, `encryptedJson()`, `timestampWithTimeZone()`, `timestamps()`, `lockVersion()`, and `uuidV7PrimaryKey()` from `src/db/lib/columns.server.ts`; do not inline equivalent Drizzle builders.
   - Drizzle Effect query failures store the Effect SQL error inside an `EffectDrizzleQueryError` Cause; inspect it with Effect's `Cause` utilities and then standard JavaScript `cause` links instead of traversing arbitrary fields or parsing error messages.
   - Keep domain schema definitions in responsibility-named `src/db/schema/*.server.ts` files and re-export every table and relation Drizzle Kit must discover from `schema.server.ts`.
   - Keep every organization-owned table in `src/db/schema/organizations.server.ts` and its relations centralized; every table must define its own `id: uuidV7PrimaryKey()`, including one-to-one, join, and third-party-managed tables.
