+++ webapp/src/db/lib/columns.server.ts
@@ -1,10 +1,46 @@
 import { sql } from "drizzle-orm"
 import { customType, integer, timestamp, uuid } from "drizzle-orm/pg-core"
 
+import {
+  type DatabaseEncryptionKeyring,
+  getDatabaseEncryptionKeyring,
+} from "./database-credentials.server.ts"
+import { decryptDatabaseValue, encryptDatabaseValue } from "./encryption.server.ts"
+
+type EncryptedJsonOptions<Value> = {
+  decode: (value: unknown) => Value
+  keyring?: DatabaseEncryptionKeyring
+}
+
 // The migration installs PostgreSQL's trusted citext extension before creating these columns. https://www.postgresql.org/docs/current/citext.html
 export const caseInsensitiveText = customType<{ data: string }>({
   dataType: () => "citext",
 })
+
+/**
+ * Creates a transparent encrypted text column.
+ * Include dynamic row identity in object values and compare it with sibling columns after reads.
+ */
+export function encryptedJson<Value>(options: EncryptedJsonOptions<Value>) {
+  const keyring = () => options.keyring ?? getDatabaseEncryptionKeyring()
+  const decodeStoredValue = (storedValue: string) =>
+    decryptDatabaseValue({
+      storedValue,
+      decode: options.decode,
+      keyring: keyring(),
+    }).value
+  return customType<{ data: Value; driverData: string; jsonData: string }>({
+    dataType: () => "text",
+    toDriver: (value) =>
+      encryptDatabaseValue({
+        value,
+        decode: options.decode,
+        keyring: keyring(),
+      }),
+    fromDriver: decodeStoredValue,
+    fromJson: decodeStoredValue,
+  })()
+}
 
 export function timestampWithTimeZone() {
   return timestamp({ withTimezone: true })
